Codes reconstruction
Add strainrate function to Mesh class.

Clear up the pointer functions which have flaws in memory management. The returned pointers are either declared to be static arrays or deleted outside the function once they are not needed.

Set dimension integers to be const int.

Add a system call to "preplot" for transforming ascii files to plt files once a FIELD dat file is output.