Change spanwise mean removal, add time derivative output
Store time derivative fields of U, V, W, P in intermediate variables at every step, and output them in every file writing.
The time derivatives of inner grids are updated in IDM, and of boundary grids are updated in applyBC(). The output at step 0 is not correct since initiation only covers U, V, W, P.

Change the removal of spanwise mean from the velocity RHS to the velocities themselves after every time evolution. Since the former fails to diffuse the distortions in the mean profile.

Change the removal of spanwise mean from spectral space to the physical space, for a lower time complexity and a looser coupling of codes.

Tested on a MFU 1000 case.  Running state is ok, but the output data have not been checked.