Add OMP, add post process codes
Add OpenMP feature to the program.
The omp parallelisation is only applied to uhcalc() of IDM class, since it is the most time consuming part.
Other parts of the program remain to be serial so that no  consideration of parallelisation is needed out of IDM.
Number of threads is detrmined automatically by the total grid number.
A timer main file is also added for time tests. Note that clock() function cannot be used because it adds up all time consumed all CPUs, instead of giving out the real time.

Makefile rearranged to a clearer structure.

Three basic classes are added to the python post process codes. And a main.py module is used to generate statistics dat files.