Change TriFilter
TriFlt coefficients were changed to (1/4,1/2,1/4) in the last commit, it turned out to severely damage the mean wall stress. The reason may be that it corresponds to a filter scale (Delta) that is not equal to the grid size.

Anyways, it is changed back to (1/6,2/3,1/6) now and the result, though not perfect, is acceptable. (4% error for Ret 550)