Merge OFW codes with master
The off-wall codes differ with the master only in main.cpp, so main_ofw.cpp is added into the master, and which main file to be linked by the compiler is specified by the makefile.

For the post process codes, fixed a bug in computing the disspation on the boundary.