Add DA class
Impemented the Data Assimilation algorithm in a class named DA. Including the solver of the adjoint equations, an open interface to acquire experiment data, etc...

Modified the main function and the solver class to accommodate the DA structure.

Tested DA function on a Ret 180 case. The case ran 2000 time steps and the residual drops as expected. More tests need to be carried out.

Moved the methods of class Solver from the header file to the cpp file.

Added a 'module' method in class Vctr to compute the vector module at cell centers.

Fixed a bug in class Para. Avoid incorrect path string when an empty string is prescribed in XINDAT.