Test of DA class
Tested DA class in 2 cases with different parameters including the exit error, the max iteration number and the sampling rate of experiment data.

Tests showed that the module works well with appropriate parameters. Two flows with the same configuration but different initial conditions can be assimilated.

Modifications was made, including: normalize the error by the mean of mask function, change the way to triger iteration.