Add hybrid offwall BC with DA
Add DA algorithm to the implementation of off-wall BC. Enfore multiple layers along side the boundary to be identitcal to the full-sized channel. Tests are made against the number of assimilated layers, and results show that even with 10 assimilated layers (total 37) each boundary the discrepency grows eventually.

Some wrappings are made to the Solver class.

The adjustment of mean pressure gradient will be isolated from the IDM class, and this is a backup commit before code implementation.