Clarify the definition of CC boundaries
Cell-center boundaries are defined to align with the UW boundaries, namely the virtual boundary. This means not only that P aligns in wall-normal direction with UW, but also that when interpolating V to CC, it should align with P. Specifically, V is linearly extrapolated to boundary, which means dV/dy is nearest interpolated. This is for unified way to process boundaries in post-process.

DA and OFW in main file are merged into one main function. In the future they should be further abstracted to account for all auxiliary-main flow configurations.

Change 'char*' parameters in function parameter lists to 'const char*' so that constant strings can be passed without warning.