shear strain and shear sgs stress
Add 'SubGridShearStress' function in SGS to solve sub-grid shear stress on edges. (with dirty trick, only those at boundaries are actually solved)

Add 'ShearStrain' function in Vctr to solve shear strain rate on edges. This is frequently used in different parts of the program because it is physically important. So it is batter to gather them up instead of realizing them separately.