energy flux rescale
The energy flux filtered from DNS field need to be rescaled before supply to the off-wall simulation. Such a rescaling is chosen to be constraining R12 on the boundary face where viscosity is extrapolated from inner.

Besides, the way is found to have no effect to supply tau12, tau23 at boundary (with R12 constrained) in conjunction with energy flux at cell-center of the boundary cell.

Add threadprivate to the function solving strainrate to enable parallel.

When solving eddy viscosity, both the lower and the upper limit must be set, otherwise there would be unormal jump in total energy.