bug of flowrate fixed
Bug description: even with R12 constrained, the flowrate still goes up when supply SGS stress on the boundary.
Turned out that limiting boundary eddy viscosity to be strictly positive severely alters mean tau12 on the boundary. So it is necessary to keep negative EV.

Reorganize the parallel structure of SGS by adding threadprivate to Vector operators.