Add copy constructor and fftx to Scla
Add proper copy constructor and operator= to class Scla so that it can be passed by function return or function arguments. Classes Vctr and Flow are modified correspondingly. The orginal operator= is changed to be a function, namely Set().

Add fftx() to Scla. And change to range of fft operations to be the whole bulk including boundaries. This is for those applications (like boundary prediction) which require spectral filter on boundaries.