Fix bugs in PIO model
The indices initiated in Geometry class should have range [0 ~ Ny], not [0 ~ Ny-1], because boundary shear stress requires correct indexing of neighbour layers of the boundary.

The WallRscl(y, r) function should not constrain y to be greater or smaller than 1, because sometimes it is used to infer the inner-scaled y.

A test of uniform boundary SGS stress is also carried for Ret 540 case.