Reorganize non-dimensionalization in post processing
Move characteristic scales from Statis to DataSetInfo, since it only requires information from LOG.dat.

Add a dimconvertor module to convert the dimension of a whole dataset.

Modify the output of Euu.bin etc. to serve for linear NS optimization.

Remark: all output of postprocessing codes should be of direct use. No middle file should be saved because 1. it is difficult to make a generalized protocol for file IO; 2. it is not so costly to process raw data repeatly.