Add a compatible BC and adjust Vreman coefficient
A compatible BC is implement to allleviate the boundary ossilation in off-wall LES. It takes a portion of V from the last time step to enhance the compatibility of divergence-free condition at the current time step. However, test shows that this way can only effectively alleviate boundary ossilation when the blender factor is 1,  which means boundary V does not evolve during the whole time.

The model coefficient of Vreman SGS model is limited in a more proper manner. It recovers to 0.07 when negative values show up. So that the computation will not be stuck at negative coefficients.