IDM generalized
Generalize the IDM module to be applicable to both channel and TBL flows. The 0-indices of coordinates are used to indicate whether the direction is periodic or not.

To keep the form concise, some points that need not compute are computed anyway, which may cause nan to spread in the solution. Therefore, a new method called 'CleanBoundary' is added to the Flow class.