Bcond for ETBL
The boundary conditions for equivalent TBL was originally the same to that of half-channel. But This BC does not support flipped inlet for interrupting spurious periodicity. Therefore a new BC which is similar to that for developing TBL but with top boundary using free-slip, is added. The inlet is flipped from 2/3 streamwise position.