Fix outer-scaled flow by adjusting Fx^+
Replace the input Fx by the universal Fx^+, which requires being scaled by U_\infty^+ if the computation is non-dimensionalized by outer scales.

A technique similar to that in flowrate-fixed channel is applied to correct mpg (act as scale coefficient of Fx^+) and velocity, keeping the mean velocity at \eta=1 to be constantly 0.99.

This is tested for Re = 11168 and 10500. The results have not converged yet, but appear to be satisfying.