Modify fft_r2r plans initiation of Scla
The initiation of fft_r2r plans is a (Ny+1)*(Nz+1) loop for every Scla, and it takes place at the beginning of program where OMP has not been set up. This costs a big amount of time, and is not necessary. It is changed to use the 'fft_temp' array for fft_r2r and only requires (Ny+1) initiations.

A bug in initiating mpg is fixed by change the order of initiating mpg and ContinueCase().