Change x-BC from periodic to recycling
The periodic BC in x direction is changed to recycling, to incorporate the mirroring of inlet to interrupt spurious periodicity.

The recycling BC has two major features: 1. the mean streamwise velocity of inlet is not recycled, but evolved following the mean momentum balance; 2. the outlet flowrate is corrected to match the flowrate of the inlet.

The inner scale (mpg1) is still dynamically computed by fixing the characteristic outer velocity, and the change in mpg1 is little enough to decouple the inner scale adjustment and the inlet mean velocity evolution.