Add BC for developing BL
Modify the inlet BC for EBL: the mean velocity of the inflow (which evolves to balance the driving force and the domain-mean velocity) is directly equated to the domain-mean velocity, instead of the local mean. So the velocity increment is W.R.T the domain mean, and imposed on the inlet.

Add the option of Z-shifting to the inlet of EBL.

Add BC for developing TBL. The mean velocity of inflow is read from external file, and fluctuations are taken from the auxiliary EBL.

Tested for several EBL-TBL cases, performs well.